Fixed “ROI offsetX” and “ROI offsetY” values wraparound

+++ dct_widgets/com_ctrl/CamItf.cpp
@@ -116,7 +116,7 @@ void CamItf::GetCameraRoiOffsetInfo()
 
         // get camera info structure from device
         int res = ctrl_protocol_get_cam_roi_offset_info( GET_PROTOCOL_INSTANCE(this),
-                    GET_CHANNEL_INSTANCE(this), sizeof(i), (uint8_t *)&i );
+                    GET_CHANNEL_INSTANCE(this), sizeof(i), (uint16_t *)&i );
         HANDLE_ERROR( res );
 
         // emit a CameraRoiOffsetInfoChanged signal
@@ -137,7 +137,7 @@ void CamItf::GetCameraRoiOffset()
 
         // get camera info structure from device
         int res = ctrl_protocol_get_cam_roi_offset( GET_PROTOCOL_INSTANCE(this),
-                    GET_CHANNEL_INSTANCE(this), sizeof(i), (uint8_t *)&i );
+                    GET_CHANNEL_INSTANCE(this), sizeof(i), (uint16_t *)&i );
         HANDLE_ERROR( res );
 
         // emit a CameraRoiOffsetChanged signal
@@ -152,9 +152,9 @@ void CamItf::GetCameraRoiOffset()
 void CamItf::onCameraRoiOffsetChange( int offset_x, int offset_y )
 {
     // convert to array
-    uint8_t values[2];
-    values[0] = (uint8_t)offset_x;
-    values[1] = (uint8_t)offset_y;
+    uint16_t values[2];
+    values[0] = (uint16_t)offset_x;
+    values[1] = (uint16_t)offset_y;
 
     int res = ctrl_protocol_set_cam_roi_offset( GET_PROTOCOL_INSTANCE(this),
             GET_CHANNEL_INSTANCE(this), 2, values );
